Added addition and subtraction and default constructor fixes
Addidion and subtraction have now 1 test each and seem to work.
Specifically, subtraction has been added and for that the implementation
of `integer` changed.

The default constructor allowed the construction of "zeros" even though
the range did not include 0. This has now been fixed and helped fixing a
test bug.