Added type traits (issue #6)
Adds `includes` which checks if an element is in a range,
`includes_positives` which checks if any positive number is in the
range, `includes_negatives` which checks if any negative number is in
the range.

All three type traits work on generic "ranges", that for this library
means that `lbound`, `data_t` and `rbound` are compile-time variables
defining respectively: the left/lower bound of the range, with type
`data_t`, the type of integer this range class represents, the
right/upper bound of the range, with type `data_t`.

Adds related tests.